genai/requirements.txt
flask
flask-cors
requests
python-dotenv
pdfkit
wkhtmltopdf